requirements.txt
# Core
Django<1.11
django-model-utils==2.6
django-filter==1.0.1
djangorestframework<3.6
pycurl
psycopg2
redis
requests==2.11.1
requests_toolbelt==0.7.0
uwsgi
python-dateutil

# Testing and code quality
mock
pep8
pylint
